backend/requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.2
setuptools==70.1.1
wheel==0.43.0
